% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Babylon.R
\docType{data}
\name{Babylon}
\alias{Babylon}
\title{Create a new Babylon server manager}
\description{
The Babylon generator creates a new 'Babylon'-object, which contains logic to interact
with a running bbq server from the babylon ecosystem.
}
\section{Initialization}{

A new Babylon object is initialized with the new() method:

\code{bbq <- Babylon$new()}

and can optionally take some initialization arguments:
\itemize{
\item \code{init} - whether to initialize a bbq instance on the local machine
\item \code{host} - the hostname the server is running on - defaults to 'http://localhost'
\item \code{port} - port the server is listening on - defaults to 3333.
\item \code{verbose} - whether internal status information should be displayed
\item \code{mustWork} - on initialization, check and confirm a bbq server is listening on the host/port configuration set
}
}

\section{Methods}{

\itemize{
\item \code{get_models(status, STATUSES)} - get information about models that have been submitted
\item \code{get_model(id)} - get information about a model
}
}

\keyword{datasets}
