% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Babylon.R, R/bbq_client.R
\docType{class}
\name{Babylon}
\alias{Babylon}
\alias{bbq_client}
\title{Create a new Babylon server manager}
\usage{
bbq_client(init = NULL, host = "http://localhost", port = 3333,
  verbose = TRUE, must_work = TRUE)
}
\arguments{
\item{init}{whether to initialize bbq}

\item{host}{hostname}

\item{port}{port bbq listens on}

\item{verbose}{more information about whats going on under the hood}

\item{must_work}{must be able to ping a valid bbq server to initialize}
}
\value{
Babylon instance
}
\description{
The Babylon generator creates a new 'Babylon'-object, which contains logic to interact
with a running bbq server from the babylon ecosystem.

create a new client for babylon queue
}
\section{Initialization}{

A new Babylon object is initialized with the new() method:

\code{bbq <- Babylon$new()}

and can optionally take some initialization arguments:
\itemize{
\item \code{init} - whether to initialize a bbq instance on the local machine
\item \code{host} - the hostname the server is running on - defaults to 'http://localhost'
\item \code{port} - port the server is listening on - defaults to 3333.
\item \code{verbose} - whether internal status information should be displayed
\item \code{mustWork} - on initialization, check and confirm a bbq server is listening on the host/port configuration set
}
}

\section{Methods}{

\itemize{
\item \code{submit_models(modelpath, ...)} - check ?submit_models for arguments, the server is automatically set internally
\item \code{get_models(status, STATUSES)} - get information about models that have been submitted
\item \code{get_model(id)} - get information about a model
\item \code{poll(.ids, until, timeout, interval, print, parse)} - poll a vector of models by ID name until completion
\itemize{
\item ids - vector if model ids
\item until - status criteria to poll against, default to COMPLETED or ERROR
\item timeout - length of time to poll before stopping
\item print - whether to print the status of all models each poll
\item parse - whether to parse the http response
}
}
}

\examples{
\dontrun{
bbq <- Babylon$new()
bbq$get_models()

# get all ids
bbq$get_models \%>\% map_dbl("ID")

# find all queued models
bbq$get_models(status = "QUEUED") \%>\% map_dbl("ID")

bbq$poll(1:5) # poll for models 1-5 to complete

# get all directories for completed runs
bbq$get_models(status = "COMPLETED") \%>\% map_chr(~ .x$RunInfo$RunDir)
}
}
\keyword{datasets}
