% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_bbq.R
\name{bbq_version}
\alias{bbq_version}
\title{get the bbq version on the system}
\usage{
bbq_version(bbq_path = "bbq")
}
\arguments{
\item{bbq_path}{path to bbq}
}
\description{
get the bbq version on the system
}
