% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_bbq.R
\name{create_config}
\alias{create_config}
\title{create config}
\usage{
create_config(nm_executable = "nmfe74", clean_lvl = 2, copy_lvl = 2, ...)
}
\arguments{
\item{nm_executable}{nonmem executable}

\item{clean_lvl}{clean level 1-5}

\item{copy_lvl}{copy level 1-5}

\item{...}{additional config key values to create}
}
\description{
create config
}
