% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_est_dir_factory.R
\name{last_est_dir_factory}
\alias{last_est_dir_factory}
\title{factory to create a mapping to an estimation directory}
\usage{
last_est_dir_factory(basepath)
}
\arguments{
\item{basepath}{the base directory path of the estimation directory}
}
\description{
factory to create a mapping to an estimation directory
}
\details{
returns a function that will determine the last estimation directory
given a run name, for example, given run001_est_01 run001_est02
last_est_dir("run001") --> "path/to/run001_est_01"
}
