% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fort_factory.R
\name{read_fort_factory}
\alias{read_fort_factory}
\title{read fort files and apply a name structure}
\usage{
read_fort_factory(basepath, .fort, .names = c("ITERATION", "ID", "TIME", "F",
  "DV", "CL", "V"))
}
\arguments{
\item{basepath}{base directory}

\item{.fort}{fortran file name}

\item{.names}{column names}
}
\description{
read fort files and apply a name structure
}
\examples{
\dontrun{
read_fort51 <- read_fort_factory("../modeling", "fort.51")
read_fort51(last_est_dir("run001"))
read_fort51("run001_est_01")
}
}
