% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_bbq.R
\name{start_bbq}
\alias{start_bbq}
\title{start a bbq server}
\usage{
start_bbq(workers = parallel::detectCores(logical = FALSE),
  stdout = "bbq_server.stdout", stderr = "bbq_server.stderr")
}
\arguments{
\item{workers}{number of worker threads, defaults to number of physical cores}

\item{stdout}{where to pipe stdout}

\item{stderr}{where to pipe stderr}
}
\description{
start a bbq server
}
