% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_bbq.R
\name{start_bbq_with_config}
\alias{start_bbq_with_config}
\title{start a bbq server}
\usage{
start_bbq_with_config(config, ...)
}
\arguments{
\item{config}{list of configuration information}

\item{...}{parameters passed to \link{start_bbq}}
}
\value{
processx process
}
\description{
start a bbq server
}
\details{
if config contains a configuration labelled start_dir
will automatically start from that directory

configuration is set by writing the babylonconfig.toml
file in that directory, which should be picked up
}
\examples{
\dontrun{
bbq_pid <- start_bbq_with_config(create_config(start_dir = "path/to/dir"))
}
}
