% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_model.R
\name{submit_model}
\alias{submit_model}
\title{submit a model to an nmu server}
\usage{
submit_model(srvr, modelpath, clean_lvl = 1, copy_lvl = 1, ...,
  .run_settings = NULL, .print = FALSE, .no_submit = FALSE)
}
\arguments{
\item{srvr}{address to the server}

\item{modelpath}{absolute path to the model file to run}

\item{clean_lvl}{level to clean nonmem intermediate files}

\item{copy_lvl}{level to determine how many files to copy back to parent dir}

\item{...}{nothing right now}

\item{.run_settings}{override run settings by providing a named list}

\item{.print}{print out the json representation of settings submitted}

\item{.no_submit}{dont actually submit, instead return the list of settings}
}
\description{
submit a model to an nmu server
}
\details{
\dontrun {
# copy and clean lvl of 5 so only table files copied up and all intermediate files deleted
}
}

