% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_model.R
\name{submit_models}
\alias{submit_models}
\title{submit models to an bbq server}
\usage{
submit_models(srvr, modelpath, clean_lvl = 1, copy_lvl = 1, ...,
  .cache_dir = "", .save_exe = "", .exe_name = "", .one_est = FALSE,
  .print = FALSE, .no_submit = FALSE, .gitignore = FALSE,
  nm_executable_or_path = "nmfe74")
}
\arguments{
\item{srvr}{address to the server}

\item{modelpath}{absolute path to the model file to run}

\item{clean_lvl}{level to clean nonmem intermediate files}

\item{copy_lvl}{level to determine how many files to copy back to parent dir}

\item{...}{nothing right now}

\item{.cache_dir}{cache directory}

\item{.save_exe}{name to save the precompiled executable for later reference}

\item{.exe_name}{name of executable to use for execution}

\item{.one_est}{only estimate if no estimation dir exists}

\item{.print}{print out the json representation of settings submitted}

\item{.no_submit}{dont actually submit, instead return the list of settings}

\item{.gitignore}{add a blanket gitignore to ignore the folder from git}

\item{nm_executable_or_path}{name or path to nonmem executable, DEFAULT: nmfe74}
}
\description{
submit models to an bbq server
}
\examples{
\dontrun{
# copy and clean lvl of 5 so only table files copied up and all intermediate files deleted
res <- submit_model("http://localhost:3333/models", files, 5, 5)
}
}
